requirements.txt
hikari
hikari-lightbulb
pymongo